+++ src/game/framework/Position.java
@@ -4,6 +4,6 @@ package game.framework;
  * Created by fuve on 31/03/2017.
  */
 public interface Position {
-    int getX();
-    int getY();
+    double getX();
+    double getY();
 }
